test: [ANDROAPP-6802] move resourceManager to ui
